[feature:frontend/stock-form] - implementing use mutate stock hook
