+++ client/src/app/product/hooks/use-delete-product.tsx
@@ -13,9 +13,9 @@ export function useProductDelete() {
   const { mutateAsync: deleteProductMutation } = useMutation({
     mutationFn: async (productId: number) =>
       productService.deleteProduct(productId),
-    onSuccess: (response) => {
+    onSuccess: async (response) => {
       toast.success(response.detail);
-      queryClient.invalidateQueries({
+      await queryClient.invalidateQueries({
         queryKey: ["products", "stock-history"],
       });
     },
