+++ client/src/app/stock/components/stocks-form.tsx
@@ -1,48 +1,10 @@
 import { Modal } from "@/components/modal";
 import { BaseProps } from "@/components/base-props";
 import { ModalCloseButton } from "@/components/modal/modal-close-button";
-import { useStockProductStore } from "@/app/stock/stock.store";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { StockPayload, StockSchema } from "../stock.type";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useService } from "@/services/use-service";
-import { httpErrorHandler } from "@/lib/api";
-import { toast } from "react-toastify";
+import { useMutateStock } from "@/app/stock/hooks/use-mutate-stock";
 
 export function StocksForm({ id, className }: BaseProps) {
-  const queryClient = useQueryClient();
-  const { stockService } = useService();
-  const { product, stockType } = useStockProductStore();
-
-  const form = useForm<StockPayload>({
-    resolver: zodResolver(StockSchema),
-  });
-
-  const { mutateAsync: changeStockMutation } = useMutation({
-    mutationFn: async (data: StockPayload) => {
-      if (stockType == "INPUT") {
-        return await stockService.registerStockIn(data);
-      }
-
-      return await stockService.registerStockOut(data);
-    },
-    onSuccess: (response) => {
-      toast.success(response.detail);
-      queryClient.invalidateQueries({
-        queryKey: ["products", "stock-history"],
-      });
-    },
-    onError: httpErrorHandler,
-  });
-
-  const handleMutateStock = (data: StockPayload) => {
-    changeStockMutation(data);
-  };
-
-  if (product) {
-    form.setValue("productId", product.id);
-  }
+  const { form, product, stockType, handleMutateStock } = useMutateStock();
 
   return (
     <Modal
